Flatten project structure to root; update dependencies and configs for Next 15, Tailwind 4; migrate app from nested nexvision-ui/; remove obsolete files; adjust image domains and env; ensure build works at repo root.
Why: simplify deployment and local dev, fix missing module/build issues from nested directory and align tooling.
