+++ next.config.js
@@ -2,6 +2,12 @@
 const nextConfig = {
   images: {
     domains: ['replicate.delivery'], // Allow images from Replicate
+    remotePatterns: [
+      {
+        protocol: 'https',
+        hostname: 'replicate.delivery',
+      },
+    ],
   },
   env: {
     REPLICATE_API_TOKEN: process.env.REPLICATE_API_TOKEN,
