--- nexvision-ui/next.config.js
@@ -1,17 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  images: {
-    domains: ['replicate.delivery'], // Allow images from Replicate
-    remotePatterns: [
-      {
-        protocol: 'https',
-        hostname: 'replicate.delivery',
-      },
-    ],
-  },
-  env: {
-    REPLICATE_API_TOKEN: process.env.REPLICATE_API_TOKEN,
-  },
-}
-
-module.exports = nextConfig
