--- nexvision-ui/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "nexvision-ui",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev --turbopack",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "heic2any": "^0.0.4",
-    "next": "15.3.3",
-    "react": "^19.0.0",
-    "react-compare-image": "^3.5.4",
-    "react-dom": "^19.0.0",
-    "replicate": "^1.0.1"
-  },
-  "devDependencies": {
-    "@eslint/eslintrc": "^3",
-    "@tailwindcss/postcss": "^4",
-    "@types/node": "^20",
-    "@types/react": "^19",
-    "@types/react-dom": "^19",
-    "eslint": "^9",
-    "eslint-config-next": "15.3.3",
-    "tailwindcss": "^4",
-    "typescript": "^5"
-  }
-}
