--- nexvision-ui/src/app/globals.css
@@ -1,47 +1,0 @@
-@import "tailwindcss";
-
-:root {
-  --background: #ffffff;
-  --foreground: #171717;
-}
-
-@theme inline {
-  --color-background: var(--background);
-  --color-foreground: var(--foreground);
-  --font-sans: 'Inter', Arial, Helvetica, sans-serif;
-  --font-mono: 'Menlo', 'Monaco', 'Consolas', 'Liberation Mono', 'Courier New', monospace;
-}
-
-@media (prefers-color-scheme: dark) {
-  :root {
-    --background: #0a0a0a;
-    --foreground: #ededed;
-  }
-}
-
-body {
-  background: var(--background);
-  color: var(--foreground);
-  font-family: Arial, Helvetica, sans-serif;
-}
-
-/* Image Comparison Slider Labels */
-.react-compare-image-label {
-  font-size: 1.125rem !important;
-  font-weight: 600 !important;
-  color: white !important;
-  text-shadow: 0 2px 4px rgba(0,0,0,0.5) !important;
-  background: rgba(0,0,0,0.3) !important;
-  padding: 8px 12px !important;
-  border-radius: 8px !important;
-  margin: 1rem !important;
-}
-
-/* Ensure the slider container maintains aspect ratio */
-.react-compare-image-wrapper {
-  height: 100% !important;
-}
-
-.react-compare-image {
-  height: 100% !important;
-}
