--- nexvision-ui/src/app/layout.tsx
@@ -1,29 +1,0 @@
-import type { Metadata } from "next";
-import { Inter } from "next/font/google";
-import "./globals.css";
-import Navigation from "./components/Navigation";
-
-const inter = Inter({
-  variable: "--font-inter",
-  subsets: ["latin"],
-});
-
-export const metadata: Metadata = {
-  title: "NexVision - Reimagine Your Home",
-  description: "Transform your home's appearance instantly with AI-powered reimagining",
-};
-
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <html lang="en">
-      <body className={`${inter.variable} antialiased`}>
-        <Navigation />
-        {children}
-      </body>
-    </html>
-  );
-}
