+++ src/app/components/SafeSlider.tsx
