--- src/utils/image.ts
@@ -1,93 +1,0 @@
-export const MAX_IMAGE_SIZE = 5 * 1024 * 1024; // 5MB
-export const ALLOWED_IMAGE_TYPES = ['image/jpeg', 'image/png', 'image/webp'];
-
-export function validateImageFile(file: File): { valid: boolean; error?: string } {
-  if (file.size > MAX_IMAGE_SIZE) {
-    return {
-      valid: false,
-      error: 'Image size must be less than 5MB',
-    };
-  }
-
-  if (!ALLOWED_IMAGE_TYPES.includes(file.type)) {
-    return {
-      valid: false,
-      error: 'Only JPEG, PNG, and WebP images are supported',
-    };
-  }
-
-  return { valid: true };
-}
-
-export function formatBytes(bytes: number): string {
-  if (bytes === 0) return '0 Bytes';
-  const k = 1024;
-  const sizes = ['Bytes', 'KB', 'MB', 'GB'];
-  const i = Math.floor(Math.log(bytes) / Math.log(k));
-  return parseFloat((bytes / Math.pow(k, i)).toFixed(2)) + ' ' + sizes[i];
-}
-
-export async function compressImage(file: File, maxSizeInMB: number = 5): Promise<File> {
-  return new Promise((resolve, reject) => {
-    const reader = new FileReader();
-    reader.readAsDataURL(file);
-    reader.onload = (event) => {
-      const img = new Image();
-      img.src = event.target?.result as string;
-      img.onload = () => {
-        const canvas = document.createElement('canvas');
-        let width = img.width;
-        let height = img.height;
-
-        // Calculate new dimensions while maintaining aspect ratio
-        const maxDimension = 2048; // Max dimension for either width or height
-        if (width > height && width > maxDimension) {
-          height *= maxDimension / width;
-          width = maxDimension;
-        } else if (height > maxDimension) {
-          width *= maxDimension / height;
-          height = maxDimension;
-        }
-
-        canvas.width = width;
-        canvas.height = height;
-        const ctx = canvas.getContext('2d');
-        ctx?.drawImage(img, 0, 0, width, height);
-
-        let minQuality = 0.1;
-        let maxQuality = 0.9;
-        let quality = maxQuality;
-        let base64 = canvas.toDataURL(file.type, quality);
-        const targetSize = maxSizeInMB * 1024 * 1024 * 1.33;
-        
-        let iterations = 0;
-        while (Math.abs(base64.length - targetSize) > targetSize * 0.05 && 
-               maxQuality - minQuality > 0.01 && 
-               iterations < 10) {
-          if (base64.length > targetSize) {
-            maxQuality = quality;
-          } else {
-            minQuality = quality;
-          }
-          quality = (minQuality + maxQuality) / 2;
-          base64 = canvas.toDataURL(file.type, quality);
-          iterations++;
-        }
-
-        // Convert base64 to File
-        fetch(base64)
-          .then(res => res.blob())
-          .then(blob => {
-            const compressedFile = new File([blob], file.name, {
-              type: file.type,
-              lastModified: Date.now(),
-            });
-            resolve(compressedFile);
-          })
-          .catch(reject);
-      };
-      img.onerror = reject;
-    };
-    reader.onerror = reject;
-  });
-}
